v1_cloud_devops_repositories/ibkdataanalytics-sdl-datahub-datapipeline-1d1383e9fdc4/ibkdataanalytics-sdl-datahub-datapipeline-1d1383e9fdc4/requirements.txt
aws-cdk-lib==2.132.0
constructs>=10.0.0,<11.0.0